==> python_All_data-NA_SA/9c_ML_Export_Data_MEAN.ipynb <==
import csv
import numpy as np
import pandas as pd


# For South America

for mineral in range(0,231,10):
    age_min = str(mineral)
    met_age_min = mineral 
    met_age_max = mineral+11
    SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Neural_Network.csv')
#     SA_candidates_prob = SA_candidates_prob.loc[SA_candidates_prob['age'] >= met_age_min]
#     print(SA_candidates_prob.shape)
#     SA_candidates_prob = SA_candidates_prob.loc[SA_candidates_prob['age'] < met_age_max]
#     print(SA_candidates_prob.shape)
    SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
    print(SA_candidates_prob.shape)
    SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
    print(SA_candidates_prob.shape)
    SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
#    SA_candidates_RF_byAge.to_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Random_Forest_'+age_min+'_.csv', index=True)
    SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_MLP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

   #SA_candidates_RF_byAge = pd.concat([SA_candidates_RF_byAge, SA_candidates_prob])
    
#SA_candidates_RF_byAge.to_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Random_Forest_by_Age.csv', index=True)



import csv
import numpy as np
import pandas as pd


# For North America

for mineral in range(0,231,10):
    age_min = str(mineral)
    met_age_min = mineral 
    met_age_max = mineral+11
    NA_candidates_prob = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_GaussianProcess.csv')
    NA_candidates_prob = NA_candidates_prob[NA_candidates_prob.age >= met_age_min]
    NA_candidates_prob = NA_candidates_prob[NA_candidates_prob.age < met_age_max]
    NA_candidates_prob = NA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
    NA_candidates_prob.to_csv('case_AREPS/ml_output/NA/by_age/NA_candidates_all_prob_GP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)


import pandas as pd
from pathlib import Path

source_files = sorted(Path('case_AREPS/ml_output/SA/by_age/').glob('*.csv'))

dataframes = []
for file in source_files:
    df = pd.read_csv(file) # additional arguments up to your need
    df['classifier'] = file.name
    dataframes.append(df)

SA_by_age_merged = pd.concat(dataframes)
SA_by_age_merged.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_files_merged.csv', index=True)

import pandas as pd
from pathlib import Path

source_files = sorted(Path('case_AREPS/ml_output/NA/by_age/').glob('*.csv'))

dataframes = []
for file in source_files:
    df = pd.read_csv(file) # additional arguments up to your need
    df['classifier'] = file.name
    dataframes.append(df)

NA_by_age_merged = pd.concat(dataframes)
NA_by_age_merged.to_csv('case_AREPS/ml_output/NA/by_age/NA_candidates_all_files_merged.csv', index=True)
